MEDI-CONSULT-APPLICATION
As we know in today’s rush it is difficult to take a Doctor’s appointment in Emergency 🚑. So, what to do until we get an appointment with Doctor. That’s where this App’s comes into the action. A user can input their Symptoms and can get a list of Diseases they may poses.